Revert "Disable NXP modification that produces non compliant buffers."
By using the original upstream code paths we exposed the shader bug
converting from YUY2 to RGBA. This bypasses that bug, but reintroduces
the previously encountered conversion bugs. Those can be worked around
in different ways, so reverting this for now until the shader issue is
resolved.

One workaround is to use "glcolorconvert disable-passthrough=true".
disable-passthrough is not a standard property, added by NXP.

This reverts commit e85b8bc254d5f3e294bc0a7f134ba9ef7831756b.

Change-Id: I14c37b4bb500d439c308abeb36dded2d0fa3740b