Revert "MMFMWK-7954 waylandsink: default to play video fullscreen"
This is a poor way of making anying "fullscreen". The top weston
panel still shows and video is scaled down to account for that.
Worse, the top panels existance and size is hard coded to 32 px
at the top of the screen. Revert and make room for a proper
implementation.

BUG: 142883165

This reverts commit f4d8886d4843cd9aad392cfbca81c7dbde1216b5.

Change-Id: I18bbb3f5f05bfe3f1309b0c094fbd060d73567a4